feat: del plan and buildtasks when production build
