+++ brickyard_modules/buildtask/install/package.json
@@ -11,6 +11,7 @@
   "license": "ISC",
   "devDependencies": {
     "bower": "^1.8.0",
+    "del": "^2.2.2",
     "gulp-json-editor": "^2.2.1",
     "gulp-load-plugins": "^1.2.4",
     "lodash": "^4.6.1"
