+++ brickyard_modules/plan/mocha/index.js
@@ -1,5 +1,6 @@
 module.exports = {
 	modules: [
+		'buildtask/install',
 		'buildtask/test-mocha',
 	],
 	config: {
