+++ lib/gulpfile.js
@@ -2,34 +2,8 @@ const childProcess = require('child_process')
 const _ = require('lodash')
 const argv = require('../bin')
 const brickyard = require('./brickyard')
-const Bluebird = require('bluebird')
-const runSequenceAsync = Bluebird.promisify(require('run-sequence'))
 require('./hack')
-
-const subTasks = { build: [], run: [], test: [] }
-
-require('gulp').register_sub_tasks = (type, priority, taskName) => {
-	if (!subTasks[type]) {
-		throw new Error('sub tasks type must be', _.keys(subTasks))
-	}
-
-	if (!subTasks[type][priority]) {
-		subTasks[type][priority] = taskName
-	} else {
-		subTasks[type][priority] = _.flatten([subTasks[type][priority], taskName])
-	}
-}
-
-async function runFlatten(type) {
-	if (!subTasks[type]) {
-		throw new Error('sub tasks type must be', _.keys(subTasks))
-	}
-	const tasks = subTasks[type].filter(task => !!task)
-	console.debug(`Run Flatten : [${type}] : [${tasks}]`)
-	if (tasks.length > 0) {
-		await runSequenceAsync(...tasks)
-	}
-}
+const runFlatten = require('gulp').runFlatten
 
 const gulp = {}
 gulp.myTasks = []
