+++ lib/hack.js
@@ -1,7 +1,10 @@
 const _ = require('lodash')
 const gulp = require('gulp')
+const Bluebird = require('bluebird')
 const mergeStream = require('merge-stream')
 const runSequence = require('run-sequence')
+
+const runSequenceAsync = Bluebird.promisify(runSequence)
 
 const argv = require('../bin')
 const logger = require('./logger')
@@ -15,6 +18,32 @@ gulp.create_tasks = (group) => {
 	_.each(group, (task, id) => {
 		gulp.task(id, task)
 	})
+}
+
+const subTasks = { build: [], run: [], test: [] }
+
+gulp.register_sub_tasks = (type, priority, taskName) => {
+	if (!subTasks[type]) {
+		throw new Error('sub tasks type must be', _.keys(subTasks))
+	}
+
+	if (!subTasks[type][priority]) {
+		subTasks[type][priority] = taskName
+	} else {
+		subTasks[type][priority] = _.flatten([subTasks[type][priority], taskName])
+	}
+}
+
+gulp.runFlatten = (type) => {
+	if (!subTasks[type]) {
+		throw new Error('sub tasks type must be', _.keys(subTasks))
+	}
+	const tasks = subTasks[type].filter(task => !!task)
+	console.debug(`Run Flatten : [${type}] : [${tasks}]`)
+	if (tasks.length > 0) {
+		return runSequenceAsync(...tasks)
+	}
+	return Bluebird.resolve()
 }
 
 const alias = {}
